refactor(front): Pass A misc gotchas — move DTOs + pure local types to types/api
Final Pass A batch (assorted domains).

- data_sources/check_stuck: moved PendingActivityInfo, StuckWorkflowInfo,
  CheckStuckResponseBody; lib back-imports them.
- members: pure-contract, fully moved to types/api/members.ts (lib deleted).
- oauth/providers/mcp: moved the MCP OAuth metadata schemas/type +
  DiscoverOAuthMetadataResponseBody; lib back-imports the schemas/type it extends.
- search: moved DataSourceContentNode, PostWorkspaceSearchResponseBody, and the
  search request-schema tree; lib back-imports DataSourceContentNode +
  SearchRequestBodyType.
- spaces: appended 9 DTO/shape symbols (RichSpaceType, the space request/response
  bodies, ...) to the existing types/api/spaces.ts.
- 23 consumers updated (split-aware).

Deferred: file_system/types.ts — its entry types are pure (not a Pass B coupling)
but it's a foundational module (class + helpers + ~21 consumers); warrants its own
dedicated batch.
