+++ front-api/routes/w/[wId]/spaces/check-name.ts
@@ -1,5 +1,5 @@
-import type { CheckNameResponseBody } from "@app/lib/api/spaces";
 import { SpaceResource } from "@app/lib/resources/space_resource";
+import type { CheckNameResponseBody } from "@app/types/api/spaces";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
