+++ front/components/spaces/CreateOrEditSpaceModal.tsx
@@ -3,7 +3,6 @@ import type { SearchMemberType } from "@app/components/members/MemberSelectionTa
 import { ConfirmDeleteSpaceDialog } from "@app/components/spaces/ConfirmDeleteSpaceDialog";
 import { RestrictedAccessBody } from "@app/components/spaces/RestrictedAccessBody";
 import { RestrictedAccessHeader } from "@app/components/spaces/RestrictedAccessHeader";
-import type { SpaceCategoryInfo } from "@app/lib/api/spaces";
 import { useAuth } from "@app/lib/auth/AuthContext";
 import { useAppRouter } from "@app/lib/platform";
 import { useGroups } from "@app/lib/swr/groups";
@@ -13,6 +12,7 @@ import {
   useSpaceInfo,
   useUpdateSpace,
 } from "@app/lib/swr/spaces";
+import type { SpaceCategoryInfo } from "@app/types/api/spaces";
 import type { GroupType } from "@app/types/groups";
 import type { PlanType } from "@app/types/plan";
 import type { SpaceType } from "@app/types/space";
