+++ front/lib/swr/spaces.ts
@@ -3,17 +3,6 @@ import type {
   CursorPaginationParams,
   SortingParams,
 } from "@app/lib/api/pagination";
-import type {
-  DataSourceContentNode,
-  PostWorkspaceSearchResponseBody,
-} from "@app/lib/api/search";
-import type {
-  GetSpaceResponseBody,
-  GetSpacesResponseBody,
-  PatchSpaceResponseBody,
-  PostSpaceRequestBodyType,
-  PostSpacesResponseBody,
-} from "@app/lib/api/spaces";
 import type { PatchSpaceMembersRequestBodyType } from "@app/lib/api/spaces/members";
 import { getDisplayNameForDataSource } from "@app/lib/data_sources";
 import { clientFetch } from "@app/lib/egress/client";
@@ -32,6 +21,17 @@ import type {
 import type { PostSpaceDataSourceResponseBody } from "@app/types/api/data_sources";
 import type { SpacesLookupResponseBody } from "@app/types/api/projects/list";
 import type { DataSourceViewCategoryWithoutApps } from "@app/types/api/public/spaces";
+import type {
+  DataSourceContentNode,
+  PostWorkspaceSearchResponseBody,
+} from "@app/types/api/search";
+import type {
+  GetSpaceResponseBody,
+  GetSpacesResponseBody,
+  PatchSpaceResponseBody,
+  PostSpaceRequestBodyType,
+  PostSpacesResponseBody,
+} from "@app/types/api/spaces";
 import type { ContentNodesViewType } from "@app/types/connectors/content_nodes";
 import type { SearchWarningCode } from "@app/types/core/core_api";
 import { MIN_SEARCH_QUERY_SIZE } from "@app/types/core/utils";
